+++ Tests/SentryTests/SentryOptionsTest.m
@@ -656,6 +656,7 @@
     XCTAssertEqual(options.enableUserInteractionTracing, YES);
     XCTAssertEqual(options.enablePreWarmedAppStartTracing, NO);
     XCTAssertEqual(options.attachViewHierarchy, NO);
+    XCTAssertEqual(options.reportAccessibilityIdentifier, YES);
     XCTAssertEqual(options.experimental.sessionReplay.errorSampleRate, 0);
     XCTAssertEqual(options.experimental.sessionReplay.sessionSampleRate, 0);
 #endif // SENTRY_HAS_UIKIT
@@ -803,6 +804,11 @@
 - (void)testAttachScreenshot
 {
     [self testBooleanField:@"attachScreenshot" defaultValue:NO];
+}
+
+- (void)testReportAccessibilityIdentifier
+{
+    [self testBooleanField:@"reportAccessibilityIdentifier" defaultValue:YES];
 }
 
 - (void)testEnableUserInteractionTracing